backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
cffi==1.17.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.2.4
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
